chore: refactor filenames from italian to english for uniformity
