+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/ui/parkingareas/ParkingAreasFragment.kt
@@ -1,4 +1,4 @@
-package com.contextaware.app_bpm.ui.eventipersonali
+package com.contextaware.app_bpm.ui.parkingareas
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.contextaware.app_bpm.databinding.FragmentSlideshowBinding
+import com.contextaware.app_bpm.databinding.FragmentGalleryBinding
 
-class EventiPersonaliFragment : Fragment() {
+class ParkingAreasFragment : Fragment() {
 
-    private var _binding: FragmentSlideshowBinding? = null
+    private var _binding: FragmentGalleryBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class EventiPersonaliFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val slideshowViewModel =
-            ViewModelProvider(this).get(EventiPersonaliViewModel::class.java)
+        val galleryViewModel =
+            ViewModelProvider(this).get(ParkingAreasViewModel::class.java)
 
-        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textSlideshow
-        slideshowViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textGallery
+        galleryViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
