+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/ui/parkingareas/ParkingAreasViewModel.kt
@@ -1,13 +1,13 @@
-package com.contextaware.app_bpm.ui.parcheggiami
+package com.contextaware.app_bpm.ui.parkingareas
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ParcheggiamiViewModel : ViewModel() {
+class ParkingAreasViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
