+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/ui/parkme/ParkMe.kt
@@ -1,13 +1,13 @@
-package com.contextaware.app_bpm.ui.areeparcheggio
+package com.contextaware.app_bpm.ui.parkme
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class AreeParcheggioViewModel : ViewModel() {
+class ParkMe : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
