+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/ui/personalevents/PersonalEventsFragment.kt
@@ -1,4 +1,4 @@
-package com.contextaware.app_bpm.ui.parcheggiami
+package com.contextaware.app_bpm.ui.personalevents
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.contextaware.app_bpm.databinding.FragmentHomeBinding
+import com.contextaware.app_bpm.databinding.FragmentSlideshowBinding
 
-class ParcheggiamiFragment : Fragment() {
+class PersonalEventsFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentSlideshowBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class ParcheggiamiFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val homeViewModel =
-            ViewModelProvider(this).get(ParcheggiamiViewModel::class.java)
+        val slideshowViewModel =
+            ViewModelProvider(this).get(EventiPersonaliViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textSlideshow
+        slideshowViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
